created types directory to manage the whole type
