+++ frontend/components/Task/Task.tsx
@@ -1,17 +1,10 @@
-import { TaskParam } from "../../constants/type";
 import { useSortable } from "@dnd-kit/sortable";
 import { CSS } from "@dnd-kit/utilities";
 import React, { useState } from "react";
 import EditModal from "../Modal/EditModal";
 import Image from "next/image";
 import editIcon from "../../public/img/icon-edit.svg";
-
-type TaskProps = {
-  task: TaskParam;
-  // onUpdateTask: (updatedTask: TaskParam) => void;
-  // onDeleteTask: (taskId: number) => void;
-  // onCompleteTask: (taskId: number) => void;
-};
+import { TaskParam, TaskProps } from "@/types/type";
 
 const Task: React.FC<TaskProps> = ({ task }) => {
   const id = task.id;
